FEAT: 일부 정보로 게임 생성 기능

+++ src/main/java/com/elice/boardgame/game/entity/GameRate.java
@@ -25,5 +25,5 @@ public class GameRate extends BaseEntity {
     @JoinColumn(name = "game_id", nullable = false)
     private BoardGame boardGame;
 
-    private int rate;
+    private Double rate;
 }
